requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.2.1
redis==5.0.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
gunicorn==21.2.0
PyJWT==2.8.0
pytest==8.0.2
pytest-flask==1.3.0
flask-limiter